6.0.6

+++ events/guild/messageReactionAdd.js
@@ -1,6 +1,7 @@
 const { MessageEmbed } = require("discord.js");
 const ee = require('../../botconfig/embed');
 const Schema = require("../../databases/reaction-roles");
+const Schema2 = require("../../databases/reaction-roles2");
 
 module.exports = async (client, reaction, user) => {
   let guild = client.guilds.cache.get(reaction.message.guildId)
@@ -34,6 +35,41 @@ module.exports = async (client, reaction, user) => {
         ]
       })
         .then((msg) =>{
+          try {
+            setTimeout(() => {
+              msg.delete().catch(() => {});
+            }, 6000);
+          } catch (error) {
+            console.log(error)
+          }
+        })
+    });
+  } catch (error) {
+    console.log(error)
+  }
+  try {
+    Schema2.findOne({
+      Message: reaction.message.id
+    }, async (err, data) => {
+      if (!data) return;
+      if (!Object.keys(data.RolesId).includes(reaction.emoji.name)) return;
+
+      const [roleid] = data.RolesId[reaction.emoji.name];
+
+      try {
+        await member.roles.add(roleid)
+      } catch {
+        return channel.send(`${user} You already have that role!`)
+      }
+      // return channel.send(`${user} You have obtain the <@&${roleid}> role!`)
+      return channel.send({
+          embeds: [new MessageEmbed()
+            .setColor(es.color)
+            .setFooter(client.getFooter(es))
+            .setDescription(`${user} You have obtain the <@&${roleid}> role!`)
+          ]
+        })
+        .then((msg) => {
           try {
             setTimeout(() => {
               msg.delete().catch(() => {});
